Restructure algorithm directories
