+++ visualization/app/codeCharta/util/algorithm/rectangle.ts
