+++ visualization/app/codeCharta/util/algorithm/treemap/squarifiedTreemap.ts
@@ -1,9 +1,9 @@
 import Treemap from "./treemap"
-import { CodeMapNode } from "../../codeCharta.model"
+import { CodeMapNode } from "../../../codeCharta.model"
 import Point from "../point"
-import { StreetLayoutValuedCodeMapNode } from "../streetLayoutGenerator"
+import { StreetLayoutValuedCodeMapNode } from "../../streetLayoutGenerator"
 import Rectangle from "../rectangle"
-import { StreetLayoutHelper } from "../streetLayoutHelper"
+import { StreetLayoutHelper } from "../../streetLayoutHelper"
 import HorizontalStrip from "./strip/horizontalStrip"
 import VerticalStrip from "./strip/verticalStrip"
 import Strip from "./strip/strip"
